% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asset_class_mapping}
\alias{asset_class_mapping}
\title{Mapping of names for asset classes across rate cards, vrs and invoices}
\format{A data frame with columns:
\describe{
 \item{Asset.Class}{Name on Invoices}
 \item{Adhoc.Name}{Name on Adhoc Verification Reports}
 \item{Scheduled.Name}{Name on Scheduled Verification Reports}
}}
\source{
Bloomberg invoices
}
\usage{
asset_class_mapping
}
\description{
Mapping of names for asset classes across rate cards, vrs and invoices
}
\examples{
\dontrun{
 asset_class_mapping
}
}
\keyword{datasets}
