% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_all_vr.R
\name{assign_vr_cost}
\alias{assign_vr_cost}
\title{Assign the cost of each hit according to a rate card}
\usage{
assign_vr_cost(this_usage, rate_card, conn, run = 1)
}
\arguments{
\item{this_usage}{data.frame of hits as read from DB}

\item{rate_card}{any rate card with relevant rates in effective_rate or List.Price column}

\item{conn}{DB connection}

\item{run}{integer, gives a run count to differentiate different iterations}
}
\value{
dataframe of the hits joined with vr_detail table. data to create virtual invoice also include
}
\description{
Assign the cost of each hit according to a rate card
}
