% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulk_fields}
\alias{bulk_fields}
\title{List of all Back Office fields, descriptions, and product group}
\format{A data frame with columns:}
\source{
Bloomberg
}
\usage{
bulk_fields
}
\description{
List of all Back Office fields, descriptions, and product group
}
\examples{
\dontrun{
 bulk_fields
}
}
\keyword{datasets}
