% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_all_vr.R
\name{check_db}
\alias{check_db}
\title{check_db}
\usage{
check_db(conn)
}
\arguments{
\item{conn}{A character string identifying the database location}
}
\value{
True/False. This function checks the database to see that the given entry has been processed.
}
\description{
check_db
}
