% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_all_vr.R
\name{cluster_times}
\alias{cluster_times}
\title{cluster_times}
\usage{
cluster_times(filepath)
}
\arguments{
\item{filepath}{a good description}
}
\value{
This function takes the date/time and account of each row and clusters those rows with matching info
        together to be treated as part of a single request event.
}
\description{
cluster_times
}
