% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sections}
\alias{data_sections}
\title{List of all Back Office file product file names by product group}
\format{A data frame with columns:}
\source{
Bloomberg
}
\usage{
data_sections
}
\description{
List of all Back Office file product file names by product group
}
\examples{
\dontrun{
 data_sections
}
}
\keyword{datasets}
