% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dl_cats}
\alias{dl_cats}
\title{Inventory of all Field Mnemonics available through Per Security}
\format{A data frame with columns:}
\source{
Bloomberg
}
\usage{
dl_cats
}
\description{
Inventory of all Field Mnemonics available through Per Security
}
\examples{
\dontrun{
 dl_cats
}
}
\keyword{datasets}
