% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dl_products}
\alias{dl_products}
\title{List of all Back Office offered by Bloomberg}
\format{A data frame with columns:}
\source{
Bloomberg
}
\usage{
dl_products
}
\description{
List of all Back Office offered by Bloomberg
}
\examples{
\dontrun{
 dl_products
}
}
\keyword{datasets}
