% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_all_vr.R
\name{get_vrs}
\alias{get_vrs}
\title{Get all VR files from designated directory}
\usage{
get_vrs(dir_vr)
}
\arguments{
\item{dir_vr}{the directory where we expect to find verification reports}
}
\value{
data.frame describing all verification reports found and thier attributes
}
\description{
Get all VR files from designated directory
}
\examples{
\dontrun{
  dir_vr \%>\% get_vrs()
  }

}
