% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_all_vr.R
\name{parse_all_vr}
\alias{parse_all_vr}
\title{parse_all_vr}
\usage{
parse_all_vr(conn, dir_vr, target_account, target_month)
}
\arguments{
\item{conn}{A character string identifying the database location}

\item{dir_vr}{A character string identifying the directory where verification reports are held}

\item{target_account}{The account to be parsed}

\item{target_month}{The month of verification reports to be parsed}
}
\value{
list(filepaths, # of rows)
        This function reads/cleans/writes parsed verification files to a db for later use. The function
        is designed to process the verification reports of each month of each account. Nested functions include:
        parse_detail(), parse_summary(), parse_refresh()
}
\description{
parse_all_vr
}
