% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{product_mapping}
\alias{product_mapping}
\title{Mapping of names for products across rate cards, vrs and invoices}
\format{A data frame with columns:
\describe{
 \item{Categories}{Name on Invoices}
 \item{Adhoc.Name}{Name on Adhoc Verification Reports}
 \item{Scheduled.Name}{Name on Scheduled Verification Reports}
}}
\source{
Bloomberg invoices
}
\usage{
product_mapping
}
\description{
Mapping of names for products across rate cards, vrs and invoices
}
\examples{
\dontrun{
 product_mapping
}
}
\keyword{datasets}
