% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_all_vr.R
\name{write_vr_cost}
\alias{write_vr_cost}
\title{Write the parsed VR detail file contents to DB}
\usage{
write_vr_cost(this_usage, conn, drop = F)
}
\arguments{
\item{this_usage}{dataframe of the parsed VR detail costs}

\item{conn}{DB connection SQLite object}

\item{drop}{logical if T then existing table will be dropped}
}
\value{
dataframe representing the whole table just written to DB
}
\description{
Write the parsed VR detail file contents to DB
}
