% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pub_rate_card}
\alias{pub_rate_card}
\title{Standard Enterpirse Data License Ad-Hoc rate card list prices}
\format{A data frame with columns:
\describe{
 \item{Request.Type}{Initital monthly charge (Unique) or subsequent (Access) charges.}
 \item{Data.Category}{SecMaster, Pricing, Corporate Actions, etc.}
 \item{Fee.Type}{Ad-Hoc or Scheduled, rate card model.}
 \item{Asset.Type}{Equity,Fixed Income, Muni, etc.}
 \item{Band}{Unique count range to trigger this rate}
 \item{List.Price}{Standard rate card price}
}}
\source{
Bloomberg invoices
}
\usage{
pub_rate_card
}
\description{
Standard Enterpirse Data License Ad-Hoc rate card list prices
}
\examples{
\dontrun{
 pub_rate)card
}
}
\keyword{datasets}
